requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
clip==0.2.0
clip-by-openai==1.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
ftfy==6.3.1
hf-xet==1.1.2
huggingface-hub==0.32.3
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
zipp==3.22.0
